backend/requirements.txt
fastapi>=0.135.1
uvicorn[standard]>=0.41.0
pydantic-ai-slim[openai]>=1.66
asyncpg>=0.31.0
python-multipart>=0.0.9

opentelemetry-instrumentation-fastapi>=0.61b0
opentelemetry-instrumentation-httpx>=0.61b0
opentelemetry-semantic-conventions>=0.61b0
#opentelemetry-instrumentation-openai-v2 installed from source in Dockerfile
opentelemetry-sdk>=1.40.0
opentelemetry-exporter-otlp>=1.40.0